refactor : application no longer throws error when backup is not found
