+++ src/main/java/team7/hrbank/domain/backup/service/BackupServiceImpl.java
@@ -94,5 +94,4 @@ public class BackupServiceImpl implements BackupService {
     return backupQueryService.findLatestBackupByStatus(status);
   }
 
-
 }
